change(jwt): 将 jwt 移到 utils
